pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vdiffusers"
version = "0.0.1"
description = "VDiffuser is a high-performance inference server for diffusion models with OpenAI API compatibility."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "aiohttp",
    "requests",
    "tqdm",
    "numpy",
    "IPython",
    "setproctitle",
    "blobfile==3.0.0",
    "build",
    "compressed-tensors",
    "datasets",
    "fastapi",
    "hf_transfer",
    "huggingface_hub",
    "interegular",
    "llguidance>=0.7.11,<0.8.0",
    "modelscope",
    "msgspec",
    "ninja",
    "orjson",
    "outlines==0.1.11",
    "packaging",
    "partial_json_parser",
    "pillow",
    "prometheus-client>=0.20.0",
    "psutil",
    "pydantic",
    "pynvml",
    "pybase64",
    "python-multipart",
    "pyzmq>=25.1.2",
    "sentencepiece",
    "soundfile==0.13.1",
    "scipy",
    "torchao==0.9.0",
    "transformers==4.55.0",
    "timm==1.0.16",
    "uvicorn",
    "uvloop",
    "xgrammar==0.1.22",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "torchvision",
    "cuda-python",
    "einops",
    "diffusers",
    "openai",
    "accelerate",
]

[tool.setuptools.packages.find]
exclude = [
    "assets*",
    "benchmark*",
    "docs*",
    "dist*",
    "playground*",
    "scripts*",
    "tests*",
]

[tool.wheel]
exclude = [
    "assets*",
    "benchmark*",
    "docs*",
    "dist*",
    "playground*",
    "scripts*",
    "tests*",
]

[tool.codespell]
ignore-words-list = "ans, als, hel, boostrap, childs, te, vas, hsa, ment"
skip = "*.json,*.jsonl,*.patch,*.txt"
